+++ libs/atoms.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const isNightModeAtom = atom(false);
