--- libs/recoil/isNightModeState.ts
@@ -1,8 +1,0 @@
-import { atom } from "recoil";
-
-const isNightModeState = atom({
-  key: "isNightModeState",
-  default: "",
-});
-
-export default isNightModeState;
